requirements.txt
# Heroku-optimized requirements (without heavy ML dependencies)
# Core dependencies
python-telegram-bot==21.3
google-generativeai>=0.8.0
google-genai>=1.0.0
httpx==0.28.1
python-dotenv==1.1.1
aiohttp==3.12.15
aiohttp-jinja2==1.6.0
Jinja2==3.1.5
pillow>=10.0.0
firebase-admin==6.5.0

# Document processing
PyMuPDF==1.26.4
python-docx==1.2.0
openpyxl==3.1.5
python-pptx==1.0.2
reportlab==4.4.4

# Error handling & reliability
tenacity>=8.5.0,<9.0.0